requirements.txt
astropy
astroquery
pandas
aiohttp
aiohttp_apispec
marshmallow
requests
rq
setuptools
pyyaml
click
black
isort
psycopg[binary,pool]
structlog
testcontainers-postgres
jinja2
pandas-stubs
types-pyyaml
types-psycopg2
pytest
pylint
parameterized
mypy
